pyproject.toml
[tool.poetry]
name = "do-nothing-screenplay"
version = "0.1.0"
description = ""
authors = ["Ian S Pringle <ipringle@oteemo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
PyYAML = "^6.0"
Jinja2 = "^3.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
